refactor: import main components to package's __init__.py
