+++ dequest/__init__.py
@@ -1,1 +1,22 @@
-__version__ = "0.2.0"
+from . import exceptions
+from .cache import get_cache
+from .circuit_breaker import CircuitBreaker
+from .clients import async_client, sync_client
+from .config import DequestConfig
+from .http import ConsumerType, HttpMethod
+from .parameter_types import FormParameter, JsonBody, PathParameter, QueryParameter
+
+__all__ = [
+    "async_client",
+    "sync_client",
+    "CircuitBreaker",
+    "get_cache",
+    "exceptions",
+    "ConsumerType",
+    "HttpMethod",
+    "FormParameter",
+    "JsonBody",
+    "PathParameter",
+    "QueryParameter",
+    "DequestConfig",
+]
