+++ tests/test_async_client.py
@@ -2,10 +2,7 @@ import asyncio
 
 import pytest
 
-from dequest.cache import get_cache
-from dequest.circuit_breaker import CircuitBreaker
-from dequest.clients import async_client
-from dequest.parameter_types import FormParameter, JsonBody, QueryParameter
+from dequest import CircuitBreaker, FormParameter, JsonBody, QueryParameter, async_client, get_cache
 from dequest.utils import generate_cache_key
